Refactor import statements in validator_service and views for consistency
